adding InMemStoreManager, renaming mem_kvstore -> in_memkvstore
