+++ rspace++/src/rspace/shared/mod.rs
@@ -6,4 +6,5 @@ pub mod lmdb_dir_store_manager;
 pub mod rspace_store_manager;
 mod lmdb_store_manager;
 mod lmdb_key_value_store;
-pub mod mem_key_value_store;
+pub mod in_mem_key_value_store;
+pub mod in_mem_store_manager;
