+++ rspace++/tests/export_import_tests.rs
@@ -1,0 +1,3 @@
+// See rspace/src/test/scala/coop/rchain/rspace/ExportImportTests.scala
+#[cfg(test)]
+mod tests {}
