+++ rspace++/tests/mod.rs
@@ -1,3 +1,4 @@
 mod matcher;
 mod history;
 mod storage_actions_test;
+mod export_import_tests;
